keep query syntax completely within the widget
